Made a few changes: 1) Due to running the MSVC "run code analysis" tool. 2) Due to running the AMD microperf tool. not too much and also didn't gain much perf with (2) - less than 1%   but did make the code a bit less safe so might change that back

+++ _bitutil.h
@@ -58,7 +58,7 @@ size_t MSBitSet( t_tyT _tTest )
 }
 
 // Lookup table for MSB of a nibble:
-static constexpr int v_rgiBit[16] = { 0, 0, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3 };
+static constexpr size_t v_rgiBit[16] = { 0, 0, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3 };
 
 inline
 constexpr size_t _KMSBitSet8( uint8_t _u8Test )
